+++ libs/components/src/lib/checkbox/checkbox.ts
@@ -101,10 +101,6 @@ export class Checkbox extends FormAssociatedCheckbox {
 		this.proxy.setAttribute('type', 'checkbox');
 	}
 
-	indeterminateChanged(_: boolean, next: boolean) {
-		this.checked = !next;
-	}
-
 	ariaCheckedChanged() {
 		if (this.ariaChecked === 'mixed') {
 			this.indeterminate = true;