+++ libs/components/src/lib/combobox/ui.test.ts
@@ -41,6 +41,6 @@ test('should show the component', async ({ page }: { page: Page }) => {
 	await page.waitForLoadState('networkidle');
 
 	expect(await testWrapper?.screenshot()).toMatchSnapshot(
-		'./snapshots/combobox.png'
+		'snapshots/combobox.png'
 	);
 });