+++ libs/components/src/lib/searchable-select/ui.test.ts
@@ -93,7 +93,7 @@ test('should show the component', async ({ page }: { page: Page }) => {
 	await page.waitForLoadState('networkidle');
 
 	expect(await testWrapper?.screenshot()).toMatchSnapshot(
-		'./snapshots/searchable-select.png'
+		'snapshots/searchable-select.png'
 	);
 });