+++ libs/components/src/lib/tab/ui.test.ts
@@ -33,7 +33,5 @@ test('should show the component', async ({ page }: { page: Page }) => {
 
 	await page.waitForLoadState('networkidle');
 
-	expect(await testWrapper?.screenshot()).toMatchSnapshot(
-		'./snapshots/tab.png'
-	);
+	expect(await testWrapper?.screenshot()).toMatchSnapshot('snapshots/tab.png');
 });